backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==1.10.15
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-dotenv==0.19.0
aiofiles==0.7.0
websockets==10.4
python-dateutil==2.8.2
mysql-connector-python
werkzeug
pyahocorasick
scikit-learn
numpy
Flask
paramiko==3.5.0